Make the sass-maven-plugin compatible with Windows-style paths.
In the former version of the sass-maven-plugin implementation on a
Windows system, the paths to directories wouldn't be written correctly
into the ruby script calls which led to errors when actually executing
one of the mojos.

To fix the problem this commit adds an escape mechanism that makes
Windows style paths that contain backslashes "ruby-safe" to be written
into the executed ruby script.